criação de calculadora

+++ src/br/dio/calculadora/Calculadora.java
@@ -7,7 +7,7 @@ public class Calculadora {
 	public static void main(String[] args) {
 		Scanner scan = new Scanner(System.in);
 		
-		 int a, b ;
+		 double a, b ;
 		 
 		 System.out.println("Digite o primneiro valor: ");
 		 a = scan.nextInt();
@@ -27,19 +27,19 @@ public class Calculadora {
 		 
 	}
 
-	public static int soma(int a, int b) {
+	public static double soma(double a, double b) {
 		return a + b;
 	}
 
-	public static int subtracao(int a, int b) {
+	public static double subtracao(double a, double b) {
 		return a - b;
 	}
 
-	public static int divisao(int a, int b) {
+	public static double divisao(double a, double b) {
 		return a / b;
 	}
 
-	public static int multiplicacao(int a, int b) {
+	public static double multiplicacao(double a, double b) {
 		return a * b;
 	}
